pyproject.toml
[tool.poetry]
name = "techflurry-controller"
version = "0.0.0"
description = ""
authors = ["Dániel Hagyárossy <daniel@hagyarossy.hu>","Laszlo Beres <laszloberes@hotmail.hu>"]
readme = "README.md"
license = "MIT"
# homepage = "https://"
# documentation = "https://"
keywords = ["home automation", "python", "hvac", "techflurry"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
# This is required when creating a namespaced package.
packages = [{include = "techflurry", from = "src"}]

[tool.poetry.dependencies]
python = "^3.7"
rich = "^10.16.1"
typer = "^0.4.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
# These are required packages for the CI/CD workflows
black = "^21.11b1"
coverage = "^6.1.2"
darglint = "^1.8.1"
flake8-bugbear = "^21.9.2"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.910"
pytest = "^6.2.5"
pytest-clarity = "^1.0.1"
tox = "^3.24.4"

# It is recommended to use dynamic versioning.
[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.poetry-dynamic-versioning.substitution]
files = ["src/*/__init__.py", "tests/*.py"]
patterns = ["(__version__\\s*=(?:=)?\\s*['\"])[^'\"]*(['\"])"]

[tool.black]
line-length = 79
include = '''
(
    ^/tests/*.py
  | ^/src/*
  | ^/conftest[.]py
)
'''
exclude = '''
(
    __pycache__
  | ^/tests/pytest.ini
  | ^.bankruptcy/
  | ^.venv
)
'''

[tool.isort]
extend_skip = [".bankruptcy", ".venv"]
profile = "black"
line_length = 79
multi_line_output = 3

[tool.mypy]
exclude = "(tests|.bankruptcy|.venv)"
mypy_path = "src"
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true
namespace_packages = true
explicit_package_bases = true

[build-system]
requires = ["poetry-core>=1.0.2", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"
